temp_insight_deploy/backend/requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asgiref==3.11.0
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
av==16.1.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
brotli==1.2.0
build==1.4.0
cachetools==6.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
ctranslate2==4.6.3
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.3.0
ddgs==9.10.0
debugpy==1.8.19
decorator==5.2.1
diffusers @ git+https://github.com/huggingface/diffusers.git@b712042da14d30d258d5a2e0f9f4b2a1798c4f2b
distro==1.9.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.14
duckduckgo_search==8.1.1
durationpy==0.10
eng_to_ipa==0.0.2
et_xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.13.2
fake-useragent==2.2.0
fastapi==0.128.0
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
gradio==6.5.1
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.7
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.13.0
keras-self-attention==0.51.0
kiwisolver==1.4.9
kombu==5.6.2
kubernetes==35.0.0
langchain==1.2.6
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-experimental==0.4.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langchainhub==0.1.21
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
lazy_loader==0.4
Levenshtein==0.27.3
libclang==18.1.1
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.4
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
namex==0.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.63.1
numexpr==2.14.1
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==2.15.0
openai-whisper==20250625
opencv-contrib-python==4.13.0.90
opencv-python==4.12.0.88
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
peft==0.18.1
pexpect==4.9.0
pgvector==0.4.2
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
posthog==5.4.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.5
psutil==7.2.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMatting==1.1.15
pymongo==4.16.0
PyMySQL==1.1.2
pyparsing==3.3.1
pypdf==6.6.0
PyPDF2==3.0.1
pyphen==0.17.2
PyPika==0.50.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
redis==7.1.0
referencing==0.37.0
regex==2026.1.15
rembg==2.0.69
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.2.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
streamlit==1.53.0
syllapy==0.7.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2026.1.28
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.13.3
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
typer==0.21.1
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.13.0
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.4.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.4
wikipedia==1.4.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
youtube-comment-downloader==0.1.78
zipp==3.23.0
zstandard==0.25.0
